tasks/image_classifier/requirements.txt
torch>=2.0
torchvision>=0.15
torchinfo>=1.8
wandb>=0.15
ray[tune]>=2.0
hydra-core>=1.3
omegaconf>=2.3